Test proxy-manager-bridge in a different way
Instead of using groups, we now just use `markTestAsSkipped`. We make also run the test suite when `symfony/proxy-manager-bridge` is removed.
