+++ packages/symfony-bridge/tests/Functional/SmokeTest.php
@@ -35,37 +35,6 @@ class SmokeTest extends KernelTestCase
     /**
      * @test
      */
-    public function it_handles_a_command_then_dispatches_events_for_all_modified_entities()
-    {
-        self::bootKernel(['environment' => 'config1']);
-        $container = self::$kernel->getContainer();
-
-        $this->createSchema($container);
-
-        $commandBus = $container->get('command_bus');
-        $command = new TestCommand();
-        $commandBus->handle($command);
-
-        $this->assertTrue($container->get('test_command_handler')->commandHandled);
-        $this->assertTrue($container->get('test_event_subscriber')->eventHandled);
-
-        // some_other_test_command is triggered by test_event_handler
-        $this->assertTrue($container->get('some_other_test_command_handler')->commandHandled);
-        $this->assertTrue($container->get('some_other_event_subscriber')->eventHandled);
-
-        // it has logged some things
-        $loggedMessages = file_get_contents($container->getParameter('log_file'));
-        $this->assertStringContainsString('command_bus.DEBUG: Started handling a message', $loggedMessages);
-        $this->assertStringContainsString('command_bus.DEBUG: Finished handling a message', $loggedMessages);
-        $this->assertStringContainsString('event_bus.DEBUG: Started handling a message', $loggedMessages);
-        $this->assertStringContainsString('event_bus.DEBUG: Finished handling a message', $loggedMessages);
-        $this->assertStringContainsString('event_bus.DEBUG: Started notifying a subscriber', $loggedMessages);
-        $this->assertStringContainsString('event_bus.DEBUG: Finished notifying a subscriber', $loggedMessages);
-    }
-
-    /**
-     * @test
-     */
     public function it_can_auto_register_event_subscribers_using_invoke()
     {
         self::bootKernel(['environment' => 'config2']);
@@ -133,26 +102,5 @@ class SmokeTest extends KernelTestCase
         $container->get('command_bus')->handle($command);
 
         $this->assertTrue($command->isHandled());
-    }
-
-    /**
-     * @test
-     * @group SymfonyBridgeProxyManagerDependency
-     */
-    public function fails_because_of_mising_dependency()
-    {
-        $this->expectException(LogicException::class);
-        $this->expectExceptionMessage('In order to use bundle "DoctrineOrmBridgeBundle" you need to require "symfony/proxy-manager-bridge" package.');
-
-        self::bootKernel(['environment' => 'config2']);
-    }
-
-    private function createSchema(ContainerInterface $container)
-    {
-        $entityManager = $container->get('doctrine.orm.entity_manager');
-        /** @var EntityManager $entityManager */
-
-        $schemaTool = new SchemaTool($entityManager);
-        $schemaTool->createSchema($entityManager->getMetadataFactory()->getAllMetadata());
     }
 }
